Rename module from dsd to nuad
