--- dsd/__init__.py
@@ -1,8 +1,0 @@
-import dsd.constraints
-import dsd.vienna_nupack
-import dsd.np
-import dsd.stopwatch
-import dsd.modifications
-
-from dsd.constraints import logger
-
