+++ examples/sample_designer.py
@@ -2,9 +2,9 @@ from typing import NamedTuple, Optional
 import os
 import argparse
 
-import dsd.constraints as dc  # type: ignore
-import dsd.vienna_nupack as dv  # type: ignore
-import dsd.search as ds  # type: ignore
+import nuad.constraints as dc  # type: ignore
+import nuad.vienna_nupack as dv  # type: ignore
+import nuad.search as ds  # type: ignore
 
 
 # command-line arguments
